+++ src/Screen/src/AchievementView.h
@@ -7,8 +7,10 @@
 #include "View.h"
 
 class AchievementView : public View {
+    const std::string _name {"Achievements"};
 public:
     void render() override;
+    const std::string &name() override { return _name; };
     void handleInput(const std::string &input) override;
 };