+++ src/Screen/src/OptionsView.h
@@ -7,8 +7,10 @@
 #include "View.h"
 
 class OptionsView : public View {
+    const std::string _name {"Options"};
 public:
     void render() override;
+    const std::string &name() override { return _name; };
     void handleInput(const std::string &input) override;
 };