+++ src/Screen/src/StatusView.cpp
@@ -9,21 +9,17 @@
 void StatusView::render() {
     std::cout << escapeCode.eraseCurrentLine;
     std::cout << "Cookies\t:\t";
-    if (wallet.getCookieAmount() > 0)
-        std::cout << cp.print(wallet.getCookieAmount());
+    if (wallet != nullptr && wallet->getCookieAmount() > 0)
+        std::cout << cp.print(wallet->getCookieAmount());
     std::cout << std::endl;
     std::cout << escapeCode.eraseCurrentLine;
     std::cout << "cps\t:\t";
-    if (wallet.getCps() > 0)
-        std::cout << cp.print(wallet.getCps());
+    if (wallet != nullptr && wallet->getCps() > 0)
+        std::cout << cp.print(wallet->getCps());
     std::cout << std::endl;
     std::cout << escapeCode.eraseCurrentLine;
 }
 
 void StatusView::handleInput(const std::string &input) {
-
-}
-
-StatusView::StatusView(Wallet &wallet) : wallet(wallet) {
 
 }