+++ src/Screen/src/StatusView.h
@@ -9,10 +9,12 @@
 
 class StatusView : public View {
 private:
-    Wallet& wallet;
-    
+    const Wallet* wallet;
+    const std::string _name {"Status"};
+
 public:
-    explicit StatusView(Wallet &wallet);
+    const std::string &name() override { return _name; };
+    explicit StatusView(const Wallet* wallet) : wallet(wallet) {};
     void render() override;
     void handleInput(const std::string &input) override;
 };