+++ src/Screen/tst/Screen-test.cpp
@@ -1,18 +1,20 @@
 #include <memory>
 #include "gtest/gtest.h"
-
 #define private public
-
 #include "Screen.h"
 
 struct ScreenTestSuite : public ::testing::Test {
     notifyMessage msg;
     Wallet wallet;
+    Inventory inventory;
+    std::string InventoryInputKey = "2";
+    std::string AchievementInputKey = "3";
+    std::string OptionsInputKey = "4";
 };
 
 TEST_F(ScreenTestSuite, defaultScreen) {
     //arrange
-    Screen screen(wallet, msg);
+    Screen screen(&wallet, &msg, &inventory);
     //act
     //assert
     ASSERT_EQ(screen.activeView, dynamic_cast<View *>(&screen.storeview));
@@ -20,13 +22,13 @@ TEST_F(ScreenTestSuite, defaultScreen) {
 
 TEST_F(ScreenTestSuite, switchScreen) {
     //arrange
-    Screen screen1(wallet, msg);
-    Screen screen2(wallet, msg);
-    Screen screen3(wallet, msg);
+    Screen screen1(&wallet, &msg, &inventory);
+    Screen screen2(&wallet, &msg, &inventory);
+    Screen screen3(&wallet, &msg, &inventory);
     //act
-    screen1.handleInput("3");
-    screen2.handleInput("4");
-    screen3.handleInput("5");
+    screen1.handleInput(InventoryInputKey);
+    screen2.handleInput(AchievementInputKey);
+    screen3.handleInput(OptionsInputKey);
     //assert
     ASSERT_EQ(screen1.activeView, dynamic_cast<View *>(&screen1.inventoryview));
     ASSERT_EQ(screen2.activeView, dynamic_cast<View *>(&screen2.achievementview));